Updating the tool
Currently, it generates an XSLT file with code created specifically for the user's given IDoc XML. In addition to that, the program also outputs a simplified version of the IDoc XML in a generic XML format and extracts all values into an array. This makes it easier for middleware or individuals to work with.